Change the unzip library.

+++ src/extract.ts
@@ -1,6 +1,6 @@
 "use strict";
 
-import * as AdmZip from "adm-zip";
+import * as decompress from "extract-zip";
 import * as fs from "fs";
 import * as path from "path";
 import createError from "./createError";
@@ -10,30 +10,34 @@ import createError from "./createError";
  */
 const extract = (from: string, to: string, logger: BowerLogger): Promise<any> => {
     return new Promise((resolve: Function, reject: Function) => {
-        // Validate that the source is a zip archive
+        // Validate that the source as supported extension
         if (!(/.*\.upack$/.test(from))) {
             return reject(createError(`File ${from} is not a known archive`, "ENOTARCHIVE"));
-        }
-
-        // If the src exist
-        fs.stat(from, (err, stats) => {
-            if (err) {
-                throw err;
-            } else {
-                if (stats.size <= 8) {
-                    reject(createError(`File ${from} is an invalid archive`, "ENOTARCHIVE"));
+        } else {
+            // If the src exist
+            fs.stat(from, (err, stats) => {
+                if (err) {
+                    throw err;
                 } else {
-                    // Extract archive
-                    new AdmZip(from).extractAllTo(to);
-                    logger.action("extract", "upack.json");
-
-                    // Delete the now unwanted file upack.json
-                    fs.unlink(path.join(to, "upack.json"), () => {
-                        resolve();
-                    });
+                    if (stats.size <= 8) {
+                        reject(createError(`File ${from} is an invalid archive`, "ENOTARCHIVE"));
+                    } else {
+                        // Extract archive
+                        logger.action("extract", "upack.json");
+                        decompress(from, {dir: to}, (err_) => {
+                            if (err_) {
+                                reject(err_);
+                            } else {
+                                // Delete the now unwanted file upack.json
+                                fs.unlink(path.join(to, "upack.json"), () => {
+                                    resolve();
+                                });
+                            }
+                        });
+                    }
                 }
-            }
-        });
+            });
+        }
     });
 };
